fixes suggestions
